+++ inst/shiny/A02_transformation/app.R
@@ -1,9 +1,10 @@
-#
-learndown::learndownShinyVersion("1.0.0") # Set app version
-BioDataScience::init()
+learndown::learndownShinyVersion("1.0.0")
+BioDataScience1::config()
 
 library(shiny)
 library(learndown)
+
+urchin <- data.io::read("urchin_bio", package = "data.io", lang = "fr")
 
 ui <- fluidPage(
   learndownShiny("Transformation des données : linéarisation"),
@@ -13,39 +14,33 @@ ui <- fluidPage(
       "scalex", "x",
       choices = c(
         "Aucune", "Logarithme népérien", "Racine carrée",
-        "Carré", "Exponentiel")),
+        "Carré", "Exponentielle")),
     selectInput(
       "scaley", "y",
       choices = c(
         "Aucune", "Logarithme népérien", "Racine carrée",
-        "Carré", "Exponentiel")),
+        "Carré", "Exponentielle")),
     hr(),
 
-    submitQuitButtons() # The learndown-specific buttons
+    submitQuitButtons()
   ),
 
-    mainPanel(
-      fluidRow(
-        column(width = 6,
-               plotOutput("plot")
-               ),
-        column(width = 6,
-               plotOutput("plot_transfo")
-               )
-        )
-      )
+  mainPanel(
+    fluidRow(
+      column(width = 6, plotOutput("plot")),
+      column(width = 6, plotOutput("plot_transfo"))
+    )
+  )
 )
 
 server <- function(input, output, session) {
-
-  urchin <- data.io::read("urchin_bio", package = "data.io", lang = "fr")
 
   urchin_transfo <- reactive({
 
     transformer <- function(
       x, method = c(
         "Aucune", "Logarithme népérien", "Racine carrée",
-        "Carré", "Exponentiel"), label, units) {
+        "Carré", "Exponentielle"), label, units) {
 
       method <- match.arg(method)
 
@@ -56,22 +51,21 @@ server <- function(input, output, session) {
       }
 
       data.io::labelise(switch(method,
-             "Aucune" = x,
-             "Logarithme népérien"= log(x),
-             "Racine carrée" = sqrt(x),
-             "Carré"  = x^2,
-             "Exponentiel" = exp(x)),
-             label = labs, units = units)
+        "Aucune" = x,
+        "Logarithme népérien" = log(x),
+        "Racine carrée" = sqrt(x),
+        "Carré"  = x^2,
+        "Exponentielle" = exp(x)),
+        label = labs, units = units)
     }
 
     dplyr::mutate(urchin,
-           y_transfo = transformer(height, method = input$scaley,
-                                   label = "la hauteur", units = ""),
-           x_transfo = transformer(weight, method = input$scalex,
-                                   label = "la masse", units = "")
+      y_transfo = transformer(height, method = input$scaley,
+        label = "la hauteur", units = ""),
+      x_transfo = transformer(weight, method = input$scalex,
+        label = "la masse", units = "")
     )
   })
-
 
   output$plot <- renderPlot({
     chart::chart(data = urchin, height ~  weight) +
@@ -84,17 +78,13 @@ server <- function(input, output, session) {
       ggplot2::geom_point()
   })
 
-  # This is the learndown-specific behaviour
-  # Track start, stop, inputs, errors (and possibly outputs)
-  trackEvents(session, input, output)
-  # Track the submit button and check answer
-  trackSubmit(session, input, output,
-              solution = list(scalex = "Logarithme népérien", scaley = "Logarithme népérien"),
-              comment = "transformation double-logarithmique",
-              message.success = "Correct, c'est la meilleur transformation. La transformation double-logarithmique est très souvent utilisée en biologie.",
-              message.error = "Incorrect, une meilleure transformation existe.")
-  # Track the quit button, save logs and close app after a delay (in sec)
-  trackQuit(session, input, output, delay = 60)
+  trackEvents(session, input, output, debug = TRUE)
+  trackSubmit(session, input, output, max_score = 5, solution =
+      list(scalex = "Logarithme népérien", scaley = "Logarithme népérien"),
+    comment = "transformation double-logarithmique",
+    message.success = "Correct, c'est la meilleure transformation. La transformation double-logarithmique est très souvent utilisée en biologie.",
+    message.error = "Incorrect, une meilleure transformation existe.")
+  trackQuit(session, input, output, delay = 10)
 
   }