+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changes in version 2020.0.0
 
 - Clean up. Now depends on learndown and BioDataScience only.
+- config() function added.
 
 ## Changes in version 2019.11.0